fixed errors moved BoardSize, BoardCoordinate & BoardPreciseCoordinate to ecs_board
