+++ evolvim-lib/src/lib/mod.rs
@@ -23,22 +23,17 @@ extern crate specs;
 #[macro_use]
 extern crate serde_derive;
 
-pub mod board;
 pub mod brain;
 pub mod climate;
 pub mod constants;
 pub mod ecs_board;
 pub mod neat;
-pub mod sbip;
-pub mod serde_structs;
 pub mod softbody;
 pub mod systems;
 pub mod terrain;
 pub mod time;
 
-pub use self::board::*;
 pub use self::brain::*;
 pub use self::climate::Climate;
-pub use self::sbip::*;
 pub use self::softbody::*;
 pub use self::terrain::*;
